--- src/main/java/Pau/exception/NoDescException.java
@@ -1,8 +1,0 @@
-package Pau.exception;
-
-public class NoDescException extends Exception {
-    public NoDescException(String errorMessage) {
-        System.out.println("oi write something please");
-        System.out.println(errorMessage);
-    }
-}
