--- src/main/java/Pau/exception/NoSuchTaskException.java
@@ -1,8 +1,0 @@
-package Pau.exception;
-
-public class NoSuchTaskException extends Exception{
-    public NoSuchTaskException(String errorMessage) {
-        System.out.println(errorMessage);
-        System.out.println("there is no such task!!");
-    }
-}
