+++ src/main/java/pau/task/Event.java
@@ -1,13 +1,33 @@
-package Pau.task;
+package pau.task;
 
+/**
+ * Represents an Event task that has a description of the task, a start and end date.
+ */
 public class Event extends Task {
+
+    /**
+     * Start date of an event.
+     */
     protected String start;
+
+    /**
+     * End date of an event.
+     */
     protected String end;
+
+    /**
+     * Constructs an event.
+     *
+     * @param description Description of the event.
+     * @param start Start date of the event.
+     * @param end End date of an event.
+     */
     public Event(String description, String start, String end) {
         super(description);
         this.start = start;
         this.end = end;
     }
+
     @Override
     public String writeToFile() {
         String delimiter = " | ";
