+++ src/main/java/pau/util/Storage.java
@@ -1,7 +1,7 @@
-package Pau.util;
+package pau.util;
 
-import Pau.task.Task;
-import Pau.task.TaskList;
+import pau.task.Task;
+import pau.task.TaskList;
 
 import java.io.File;
 import java.io.FileNotFoundException;
@@ -10,17 +10,32 @@ import java.io.IOException;
 import java.util.Scanner;
 
 /**
- * Represents the file used to store the list of tasks.
+ * Deals with loading tasks from the file and saving tasks in the file.
  */
 public class Storage {
-    private final String DEFAULT_STORAGE_FILEPATH;
+    
+    /**
+     * The path to the file containing tasks to be loaded.
+     */
+    private String filepath;
 
+    /**
+     * Constructs a Storage.
+     *
+     * @param filepath The path to the file containing tasks to be loaded.
+     */
     public Storage(String filepath) {
-        this.DEFAULT_STORAGE_FILEPATH = filepath;
+        this.filepath = filepath;
     }
+
+    /**
+     * Loads the tasks from the file.
+     *
+     * @return The list of tasks.
+     */
     public TaskList loadTasks() {
         try {
-            File toLoad = new File(DEFAULT_STORAGE_FILEPATH);
+            File toLoad = new File(filepath);
             Scanner scan = new Scanner(toLoad);
             TaskList list = new TaskList();
             while (scan.hasNext()) {
@@ -34,9 +49,14 @@ public class Storage {
         return new TaskList();
     }
 
+    /**
+     * Saves tasks to file.
+     *
+     * @param tasks The list of tasks to be saved.
+     */
     public void saveTasksToFile(TaskList tasks) {
         try {
-            FileWriter writer = new FileWriter(DEFAULT_STORAGE_FILEPATH);
+            FileWriter writer = new FileWriter(filepath);
             for(int i = 0; i < tasks.listSize(); i++) {
                 Task task = tasks.getTask(i);
                 writer.write(task.writeToFile() + "\n");
